+++ Tests/StreamChatTests/Query/Filter_Tests.swift
@@ -12,6 +12,11 @@ final class Filter_Tests: XCTestCase {
         XCTAssertEqual(filter.key, FilterKey<FilterTestScope, String>.testKey.rawValue)
         XCTAssertEqual(filter.value as? String, "equal value")
         XCTAssertEqual(filter.operator, FilterOperator.equal.rawValue)
+        
+        filter = .equal(.testKey, values: ["eq value 1", "eq value 2"])
+        XCTAssertEqual(filter.key, FilterKey<FilterTestScope, String>.testKey.rawValue)
+        XCTAssertEqual(filter.value as? [String], ["eq value 1", "eq value 2"])
+        XCTAssertEqual(filter.operator, FilterOperator.equal.rawValue)
 
         filter = .notEqual(.testKey, to: "not equal value")
         XCTAssertEqual(filter.key, FilterKey<FilterTestScope, String>.testKey.rawValue)
@@ -92,6 +97,11 @@ final class Filter_Tests: XCTestCase {
         // Test in filter
         filter = .init(operator: FilterOperator.in.rawValue, key: "test_key", value: [1, 2, 3], isCollectionFilter: false)
         XCTAssertEqual(filter.serialized, #"{"test_key":{"$in":[1,2,3]}}"#)
+        XCTAssertEqual(jsonString.deserializeFilter(), filter)
+        
+        // Test eq values filter
+        filter = .init(operator: FilterOperator.equal.rawValue, key: "test_key", value: [1, 2, 3], isCollectionFilter: false)
+        XCTAssertEqual(filter.serialized, #"{"test_key":{"$eq":[1,2,3]}}"#)
         XCTAssertEqual(jsonString.deserializeFilter(), filter)
 
         // Test group filter